+++ include/glow/Base/TensorSerialization.h
@@ -24,39 +24,39 @@
 
 namespace glow {
 
-/// Dump the content of \p tensor to a binary file \p filename. You can choose
-/// to dump also the tensor type in the file with the flag \p dumpType. The
-/// binary representation of data is guaranteed to preserve data precision
-/// (bit exactness) upon round-trips (dump/load).
-void dumpTensorToBinaryFile(Tensor &tensor, llvm::StringRef filename,
-                            bool dumpType = true);
+/// Tensor serialization options.
+struct TensorSerializationOptions {
+  /// Option for loading/dumping a tensor with tensor type.
+  bool withType{true};
+};
 
-/// Load the content of \p tensor from a binary file \p filename. You must
-/// specify whether the file data contains also the tensor type with the flag
-/// \p loadType such that the tensor type is set properly before the data is
-/// loaded. The binary representation of data is guaranteed to preserve data
+/// Dump the content of \p tensor to a binary file \p filename using the options
+/// \p opts. The binary representation of data is guaranteed to preserve data
 /// precision (bit exactness) upon round-trips (dump/load).
+void dumpTensorToBinaryFile(Tensor &tensor, llvm::StringRef filename,
+                            const TensorSerializationOptions &opts);
+
+/// Load the content of \p tensor from a binary file \p filename using the
+/// options \p opts. The binary representation of data is guaranteed to preserve
+/// data precision (bit exactness) upon round-trips (dump/load).
 void loadTensorFromBinaryFile(Tensor &tensor, llvm::StringRef filename,
-                              bool loadType = true);
+                              const TensorSerializationOptions &opts);
 
-/// Dump the content of \p tensor to a text file \p filename. You can choose to
-/// dump also the tensor type in the file with the flag \p dumpType. The data
-/// will be listed as a 1D array of values separated by comma (",") without
-/// other formatting. The text representation of data is NOT guaranteed to
-/// preserve data precision (bit exactness) upon round-trips (dump/load) and is
-/// used mainly for human readability.
+/// Dump the content of \p tensor to a text file \p filename using the options
+/// \p opts. The data will be listed as a 1D array of values separated by comma
+/// (",") without other formatting. The text representation of data is NOT
+/// guaranteed to preserve data precision (bit exactness) upon round-trips
+/// (dump/load) and is used mainly for human readability.
 void dumpTensorToTextFile(Tensor &tensor, llvm::StringRef filename,
-                          bool dumpType = true);
+                          const TensorSerializationOptions &opts);
 
-/// Load the content of \p tensor from a text file \p filename. You must
-/// specify whether the file data contains also the tensor type with the flag
-/// \p loadType such that the tensor type is set properly before the data is
-/// loaded. The values in the text file are expected to be listed as a 1D array
+/// Load the content of \p tensor from a text file \p filename using the options
+/// \p opts. The values in the text file are expected to be listed as a 1D array
 /// of values, separated by comma (",") without other formatting. The text
 /// representation of data is NOT guaranteed to preserve data precision (bit
 /// exactness) upon round-trips (dump/load) but is used for human readability.
 void loadTensorFromTextFile(Tensor &tensor, llvm::StringRef filename,
-                            bool loadType = true);
+                            const TensorSerializationOptions &opts);
 
 /// Load network input tensor (same one images are loaded into) from the tensor
 /// blobs files in \p filenames. All the loaded tensors are concatenated along
