+++ lib/Optimizer/IROptimizer/CMakeLists.txt
@@ -1,6 +1,8 @@
 add_library(IROptimizer
               IRFunctionPassManager.cpp
-              IROptimizer.cpp)
+              IROptimizer.cpp
+              IRInstrumentation.cpp
+              CommandLine.cpp)
 
 target_link_libraries(IROptimizer
                       PRIVATE
